Test added for subscription without machine name
This test was added to simulate the scenario where there is no machine
name, as is the case with SQS queues, as the queues just have a unique
name across the AWS account.  The subscription converter assumes that an
endpoint will be represented by both a queue and machine name.  For
example queue@machinename.